Pre Internship Task
This is a Login Page, that I have created as a part of one of the Task given by Saaspect Private Limited before the start of the actual Internship, as a part of Selection Process. I am glad, that Saaspect Private Limited has given me this opportunity to present my piece of skill to them.